Family of diagnostics functions for isotope count ratios

Cameca default CAMECA diagnostics CooksD regression diagnostics based on Cook's D

Usage

Cameca(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .alpha_level = 0.05,
  .hyp = "none",
  .mc_cores = 1
)

CooksD(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05,
  .mc_cores = 1
)

CV(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05,
  .mc_cores = 1
)

Rm(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05,
  .mc_cores = 1
)

norm_E(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05,
  .mc_cores = 1
)

QQ(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05,
  .mc_cores = 1
)

IR(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05,
  .mc_cores = 1
)

Arguments

.IC

A tibble containing processed ion count data.

.ion1

A character string constituting the heavy isotope ("13C").

.ion2

A character string constituting the light isotope ("12C").

...

Variables for grouping.

.X

A variable constituting the ion count rate (defaults to variables generated with read_IC())

.N

A variable constituting the ion counts (defaults to variables generated with read_IC().).

.species

A variable constituting the species analysed (defaults to variables generated with read_IC().).

.t

A variable constituting the time of the analyses (defaults to variables generated with read_IC().).

.output

Character string determining whether the returned values in a minimal version `"flag"` (original dataset + diagnostics) or an extended version with all the intermediate steps of ion- and isotope-wise summary statistics `"complete"`.

.alpha_level

The significance level of the hypothesis test and rejection level for outliers.

.hyp

Hypothesis test appropriate for the selected method.

.mc_cores

Number of workers for parallel execution (Does not work on Windows).

Value

A tibbletibble() containing either the original dataset with new columns related to the diagnostics or only the diagnostics. The flag variable enables convenient filtering of the original tibble for an augmentation of the original dataset.

Details

These functions perform a specific set of diagnostics to term anomalous values in raw ion count data of an isotope pair. The wrapper function diag_R is more convenient as it defines all the ion- and isotope-wise statistics required for the diagnostics.

Examples

# Descriptive an predictive statistics for 13C/12C ratios (note .output
# argument and remove zero count analysis)
tb_R <- stat_R(real_IC, "13C", "12C", file.nm, sample.nm,
               .output = "complete", .zero = TRUE)

# CAMECA style augmentation of ion count data for isotope ratios
Cameca(tb_R, "13C", "12C", file.nm, .X = Xt.pr, .N = N.pr,
       .species = species.nm, .t = t.nm, .output = "flag")
#> # A tibble: 11,700 × 25
#>    file.nm       t.nm sample.nm bl.nm.12C bl.nm.13C Xt.pr.12C Xt.pr.13C N.pr.12C
#>    <chr>        <dbl> <chr>         <int>     <int>     <dbl>     <dbl>    <dbl>
#>  1 2018-01-19-…  0.54 Belemnit…         1         1    34460.      358.    12040
#>  2 2018-01-19-…  1.08 Belemnit…         1         1    34202.      341.    11950
#>  3 2018-01-19-…  1.62 Belemnit…         1         1    34632.      395.    12100
#>  4 2018-01-19-…  2.16 Belemnit…         1         1    34191.      366.    11946
#>  5 2018-01-19-…  2.7  Belemnit…         1         1    34855.      378.    12178
#>  6 2018-01-19-…  3.24 Belemnit…         1         1    34672.      369.    12114
#>  7 2018-01-19-…  3.78 Belemnit…         1         1    34766.      369.    12147
#>  8 2018-01-19-…  4.32 Belemnit…         1         1    34609.      366.    12092
#>  9 2018-01-19-…  4.86 Belemnit…         1         1    34414.      406.    12024
#> 10 2018-01-19-…  5.4  Belemnit…         1         1    34474.      409.    12045
#> # … with 11,690 more rows, and 17 more variables: N.pr.13C <dbl>,
#> #   n_R_t.nm <int>, M_R_Xt.pr <dbl>, S_R_Xt.pr <dbl>, RS_R_Xt.pr <dbl>,
#> #   SeM_R_Xt.pr <dbl>, RSeM_R_Xt.pr <dbl>, hat_S_R_N.pr <dbl>,
#> #   hat_RS_R_N.pr <dbl>, hat_SeM_R_N.pr <dbl>, hat_RSeM_R_N.pr <dbl>,
#> #   chi2_R_N.pr <dbl>, R_Xt.pr <dbl>, ratio.nm <chr>, hat_s_R_Xt.pr <dbl>,
#> #   hat_Xt.pr.13C <dbl>, flag <fct>