Calibration of isotope ratios

calib_R function to convert isotope values back-and-forth between R and delta

Usage

calib_R(
  x,
  reference,
  isotope,
  type = "composition",
  input = "R",
  output = "F",
  y = NULL
)

Arguments

x

A numeric value or vector.

reference

A character or numeric value or vector.

isotope

A character for the name of the isotope (e.g. "13C").

type

Type of conversion, isotope composition or enrichment.

input

The type of input value, R, F, and delta.

output

Desired output value, R, F, and delta.

y

A numeric value or vector of the product when calculating enrichment factors, following the convention; x = R(product)/ y = R(substrate). Both x and y should have the same notation, either R, F or delta, as supplied to the argument input.

Value

A numeric value or vector.

Details

A fundamental in publishing isotope data is the conversion of isotope the delta formulation comparing the obtained R to that of a standard. These values are reported on a per mill scale. However in IC calculations ratios (R) or fractional abundances (F) are most commonly used. This functions provide an easy way of making these transformations for both the average composition but also fractionations as enrichment factors (alpha and epsilon).

Examples

# R value of 0.0111 to delta on VPDB scale
calib_R(0.0111, reference = 0.011237, type = "composition", input = "R",
        output = "delta")
#> [1] -12.19187
# or
calib_R(0.0111, reference = "VPDB", isotope = "13C",
        type = "composition", input = "R", output = "delta")
#> [1] -12.19187

# Fractional abundance for a given R
calib_R(0.0111, reference = 0.011237, type = "composition",
        input = "R", output = "F")
#> [1] 0.01097814

# Alpha enrichment factors can also be calculated based on two delta/R/F
# values, where the second value can be supplied as argument y (i.e.,
# the product of a reaction)
calib_R(-25, reference = "VPDB", isotope = "13C",
        type = "enrichment", input = "delta", output = "alpha", y = -105)
#> [1] 0.9179487