Process raw ion count data

cor_IC function for processing ion count data.

cor_IC(
  .IC,
  ...,
  .N = NULL,
  .t = NULL,
  .bl_t = NULL,
  .det = NULL,
  .deadtime = NULL,
  .thr_PHD = NULL,
  .M_PHD = NULL,
  .SD_PHD = NULL,
  .hide = TRUE
)

Arguments

.IC
A tibble containing raw ion count data.

...
Currently not supported.

.N
A variable constituting the ion counts.

.t
A variable constituting the time increments.

.bl_t
A variable or numeric value for the blanking time (in milliseconds).

.det
Variable or character string or variable for the detection system ("EM" or "FC").

.deadtime
A numeric value for the deadtime of the EM system with units nanoseconds.

.thr_PHD
A numeric value for the discriminator threshold of the EM. system.

.M_PHD
A variable or numeric value of the mean PHD value.

.SD_PHD
A variable or numeric value of standard deviation of the PHD value.

.hide
A logical indicating whether only processed data should be returned. If TRUE The raw data is contained as an attribute named "rawdata".

Value

A tibble::tibble() containing the original dataset and adds the variables: Xt.rw, ion count rates uncorrected for detection device-specific biases; Xt.pr, ion count rates corrected for detection device-specific biases; and N.pr, counts corrected for detection device-specific biases.

Details

The accuracy of pulsed ion counting is influenced by systematic errors which depend on the ion counting system. Deadtime and EM yield are two prominent effects for the electron multiplier systems. The deadtime refer to the timewindow when the system does not register counts; this occurs when incident ions strike the EM in a small enough time window in which the EM channel is electronically paralysed. The EM yield is the ratio between the number of output pulses counted after the EM discriminator threshold and the number of ions arriving at the EM. The latter can be gauged with the peak height distribution (PHD) which is the probability for an EM output to have a certain voltage amplitude.

Examples

# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
cor_IC(tb_rw)

#> # A tibble: 81,900 × 5
#>    file.nm                 t.nm species.nm  Xt.pr  N.pr
#>    <chr>                  <dbl> <chr>       <dbl> <dbl>
#>  1 2018-01-19-GLENDON_1_1  0.54 12C        34460. 12040
#>  2 2018-01-19-GLENDON_1_1  1.08 12C        34202. 11950
#>  3 2018-01-19-GLENDON_1_1  1.62 12C        34632. 12100
#>  4 2018-01-19-GLENDON_1_1  2.16 12C        34191. 11946
#>  5 2018-01-19-GLENDON_1_1  2.7  12C        34855. 12178
#>  6 2018-01-19-GLENDON_1_1  3.24 12C        34672. 12114
#>  7 2018-01-19-GLENDON_1_1  3.78 12C        34766. 12147
#>  8 2018-01-19-GLENDON_1_1  4.32 12C        34609. 12092
#>  9 2018-01-19-GLENDON_1_1  4.86 12C        34414. 12024
#> 10 2018-01-19-GLENDON_1_1  5.4  12C        34474. 12045
#> # … with 81,890 more rows