Create stat_R call quosure

expr_R function to generate an quosure that mimics a stat_R call for subsequent usage in dia_R

expr_R(Xt, N, species, ion1, ion2)

Arguments

Xt
A character string constituting the ion count rate.

N
A character string constituting the ion counts.

species
A character string constituting the species analysed.

ion1
A character string constituting the heavy isotope ("13C").

ion2
A character string constituting the light isotope ("12C").

Value

A list containing the input arguments as a quosure.

Details

The diag_R function performs an internal call to stat_R to perform diagnostics on the influence of individial measurements on the blockwise or global (i.e., a complete analysis) stastics. This function provides a convenient way to enter this call as an quosure into the argument args of diag_R.

Examples

expr_R(Xt = "Xt.pr", N = "N.pr", species = "species.nm", ion1 = "13C",
       ion2 = "12C")

#> <list_of<quosure>>
#> 
#> $Xt
#> <quosure>
#> expr: ^Xt.pr
#> env:  0x55c83c3f32d8
#> 
#> $N
#> <quosure>
#> expr: ^N.pr
#> env:  0x55c83c3f32d8
#> 
#> $species
#> <quosure>
#> expr: ^species.nm
#> env:  0x55c83c3f32d8
#> 
#> $ion1
#> <quosure>
#> expr: ^"13C"
#> env:  empty
#> 
#> $ion2
#> <quosure>
#> expr: ^"12C"
#> env:  empty
#>