Simulate ion count data

Simulate ion count data

Usage

simu_R(
  .sys,
  .type,
  .ion1,
  .ion2,
  .reference,
  .seed,
  .n = 3000,
  .N = 1e+06,
  .bl = 50,
  .reps = 1,
  .baseR = 0,
  .devR = 0,
  ...
)

Arguments

.sys

Systematic variation caused by ionization fluctuations as relative standard deviation of the major ion in per mille

.type

Character string to select the type of simulation "symmetric", "asymmetric" and "ideal", where the former two types introduce an offset caused by a deviation in R.

.ion1

A character string constituting the heavy isotope ("13C").

.ion2

A character string constituting the light isotope ("12C").

.reference

Character string for conversion of delta values to R (.e.g. VPDB; see ?calib_R() for more information).

.seed

Numeric sees for reproducibility of the generated data.

.n

Numeric for the number of measurements.

.N

Numeric for total ion count of the light isotope.

.bl

Numeric for block number.

.reps

Multiplication of the procedure (e.g. effectively generating multiple analyses).

.baseR

Numeric for the baseline isotope value in delta notation in per mille.

.devR

Numeric for the deviation (or forcing) away from the baseline as delta notation in per mille.

...

Not supported currently

Value

Tibble with simulated ion count data.D

Examples

# Gradient in 13C/12C over measurement transect
simu_R(50, "symmetric", "13C", "12C", "VPDB", 1)
#> # A tibble: 6,000 × 11
#>    type.nm  trend.nm base.nm force.nm  t.nm bl.nm  n.rw spot.nm species.nm  N.sm
#>    <chr>       <dbl>   <dbl>    <dbl> <int> <int> <dbl>   <int> <chr>      <dbl>
#>  1 symmetr…       50       0        0     1     1  3000       1 13C          234
#>  2 symmetr…       50       0        0     1     1  3000       1 12C        22279
#>  3 symmetr…       50       0        0     2     1  3000       1 13C          236
#>  4 symmetr…       50       0        0     2     1  3000       1 12C        22287
#>  5 symmetr…       50       0        0     3     1  3000       1 13C          286
#>  6 symmetr…       50       0        0     3     1  3000       1 12C        22234
#>  7 symmetr…       50       0        0     4     1  3000       1 13C          237
#>  8 symmetr…       50       0        0     4     1  3000       1 12C        22249
#>  9 symmetr…       50       0        0     5     1  3000       1 13C          240
#> 10 symmetr…       50       0        0     5     1  3000       1 12C        22026
#> # … with 5,990 more rows, and 1 more variable: Xt.sm <dbl>