Propagation of error for isotope ratios

stat_SDprop function for propagation of descriptive error statistics for isotope ratios (R). stat_hSDprop function for propagation of predictive (Poisson) error statistics for isotope ratios (R).

stat_SDprop(M_R, ion1, ion2, M_ion1, M_ion2, S_ion1, S_ion2, n, type = "sd")

stat_hSDprop(M_R, ion1, ion2, N_ion1, N_ion2, n, type = "sd")

Arguments

M_R
A numeric vector constituting the mean isotope ratio.

ion1
A numeric vector constituting the single ion count rate of the heavy isotope.

ion2
A numeric vector constituting the single ion count rate of the light isotope

M_ion1
A numeric vector constituting the mean ion count of the heavy isotope.

M_ion2
A numeric vector constituting the mean ion count of the light isotope.

S_ion1
A numeric vector constituting the standard deviation of the ion counts of the heavy isotope.

S_ion2
A numeric vector constituting the standard deviation of the ion counts of the light isotope.

n
A numeric vector constituting the length of the count block n.

type
A character string for the type of uncertainty estimate: "sd", standard deviation; "rsd", relative standard deviation in per mille; "se", standard error of the mean; "rse", relative standard error of the mean in per mille.

N_ion1
A numeric vector constituting the single ion counts of the heavy isotope

N_ion2
A numeric vector constituting the single ion counts of the light isotope.

Value

A numeric vector containing the propagated uncertainty of the isotope ratio.

Details

Isotope ratios are based on two measured variables, i.e., two isotopes of a single chemical species. The combined variable compounds the uncertainty associated with the individual measurement. Hence error propagation is required to obtain a reliable estimate of the uncertainty associated with the compounded variable (i.e. isotope ratio).

Examples

# Light isotope count rates
`32S` <- c(22318.19, 22151.20, 22429.52, 22143.78, 22574.25, 22455.50,
         22516.73, 22414.68, 22288.50, 22327.47)

# Heavy isotope count rate
`34S` <- c(231.4838, 220.3725, 255.5584, 237.0395, 244.4471, 238.8914,
         238.8914, 237.0395, 262.9660, 264.8179)

# Mean 32S
M_32S <- mean(`32S`)

# Mean 34S
M_34S <- mean(`34S`)

# Mean R
M_R <- M_34S / M_32S

# Standard deviation 32S
S_32S <- sd(`32S`)

# Standard deviation 34S
S_34S <- sd(`34S`)

# Propagation
stat_SDprop(M_R, `34S`, `32S`, M_34S, M_32S, S_34S, S_32S, 10, "rse")

#> [1] 17.95204