Read raw ion count data

read_IC() is designed to obtain the numerical data associated with ion counts. read_meta() can be used to specifically retrieve the metadate associated with ion count data analysis, thereby loading specifications related to the optics, the primary and secondary ion beams, and the mass spectrometer.

read_IC(directory, meta = TRUE, hide = TRUE)

read_meta(directory)

Arguments

directory
A path or connection to a directory containing raw ion count data txt files.

meta
Logical indicating whether to include metadata.

hide
Logical indicating whether metadata is included as columns FALSE or as an attribute of the tibble TRUE.

Value

A tibble::tibble containing raw ion count data and metadata.

Details

Ion count data consists of time-incremented integer values. These functions are currently only supported for data generated by a NanoSIMS50L. Raw ion count data and accompanying is extracted and collated into a single tibble from text files with the extensions .is_txt .chk_is and .stat, respectively. These files can be found in the directories associated with the SIMS measurements.

Examples

# Use point_example() to access the examples bundled with this package

read_IC(point_example("2018-01-19-GLENDON"))

#> # A tibble: 81,900 x 7
#>    file.nm                 t.nm  N.rw species.nm sample.nm          n.rw bl.nm
#>    <chr>                  <dbl> <dbl> <chr>      <chr>             <dbl> <int>
#>  1 2018-01-19-GLENDON_1_1  0.54 12040 12C        Belemnite, Indium  3900     1
#>  2 2018-01-19-GLENDON_1_1  1.08 11950 12C        Belemnite, Indium  3900     1
#>  3 2018-01-19-GLENDON_1_1  1.62 12100 12C        Belemnite, Indium  3900     1
#>  4 2018-01-19-GLENDON_1_1  2.16 11946 12C        Belemnite, Indium  3900     1
#>  5 2018-01-19-GLENDON_1_1  2.7  12178 12C        Belemnite, Indium  3900     1
#>  6 2018-01-19-GLENDON_1_1  3.24 12114 12C        Belemnite, Indium  3900     1
#>  7 2018-01-19-GLENDON_1_1  3.78 12147 12C        Belemnite, Indium  3900     1
#>  8 2018-01-19-GLENDON_1_1  4.32 12092 12C        Belemnite, Indium  3900     1
#>  9 2018-01-19-GLENDON_1_1  4.86 12024 12C        Belemnite, Indium  3900     1
#> 10 2018-01-19-GLENDON_1_1  5.4  12045 12C        Belemnite, Indium  3900     1
#> # … with 81,890 more rows